+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -87,7 +87,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-21T06:29:15.084Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-21T07:21:34.711Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")